bash: fix prefix removal
Although multiple values are provided bash might insert part of it if
the all have the same prefix. But this would skip/remove the prefix of
an ActionMultiParts as this are the display values and not what is
intended to be inserted into the line.
To prevent this a dummy value (`_`) is added.