Take advantage of ZSH _message helper
This commit takes advantage of ZSH _message helper in order to print
messages with arbitrary coloring/length, avoiding to pass the message
as a completion. There are no changes for the other shells.

Details:
The `ActionMessage` is modified, and does not wrap the message into a
completion raw value. Instead, it is stored in a variable.

Then, at the time of generating completions out of a.rawValues, we
check if there is a message in the variable. If yes:
- Non-ZSH: we make a new raw value with the message, respecting the
  context prefix and stuff, and prepend this raw value to the list of
  completion candidates. Each specific shell ActionRawValues then does
  its job as usual.
- ZSH: We pass the message as a header in the final completion string,
  and ZSH prints it with its _message helper

Advantages:
- The user cannot insert the ERR as a completion anymore
- We can print messages with arbitary coloring in them (provided the
  colors are correctly escaped, which is automatic when using packages
  such as fatih/color)
- We can print messages even when there are completion to be proposed
  besides.

Modifications:
- Added a message.go file in common package.
- Added a single call in all shells' completion Go code.
- Modified ZSH completion Go/shell code to pass the first line as a
  header, the second one with styles, and then completions.